Added ment.io to edited comments. Also updated changes to reflect new commentbox scope.
Still need to add clicky link for mentioned users.

g Changes to be committed:
